add create and retrieve endpoints
implemented offer and product classes to work with Hibernate
implemented repository interface in order to get hibernate custom
queries to work
I chose hibernate because it was in the Spring frameworks toolkit hence
I thought it would be a good idea to learn it
implemented OfferService class so that it can deal with common logical
and CRUD operations, this is where most of the logic of the application
will reside. This is because I wanted to use some of the file
structuring suggested in
https://howtodoinjava.com/spring5/webmvc/spring5-mvc-hibernate5-example/
added more assumptions re products and future improvements to this
application that I would have done if I had more time.