improved endpoint response types with appropriate status codes
got the ResponseEntity finally working like they should, realising I
should have implemented context to start with during the foundation
implementation, added that to the README future improvements section.
set up skeleton for unit tests, will work on that the next couple of
days. Thinking that saving dates as string is a bad choice but I wanted
to use LocalDate and apparently LocalDate are still new to Hibernate so
it does save them as blobs, hence, not much ordering by dates or
functionalities like that can be implemented. That could be a great
improvement in the next iterations of this API. Created a constants
class so that all the constants used on all the packages could be stored
there for ease of use.